Revert "ENH: Update neuronavigator modules"
This reverts commit 80f95c3208f09123327b436808e244413b2963a2.
