--- invesalius/.idea/dictionaries/Victor.xml
@@ -1,11 +1,0 @@
-<component name="ProjectDictionaryState">
-  <dictionary name="Victor">
-    <words>
-      <w>corregistered</w>
-      <w>corregistration</w>
-      <w>fiducial</w>
-      <w>neuronavigation</w>
-      <w>trck</w>
-    </words>
-  </dictionary>
-</component>
