+++ invesalius/data/bases.py
@@ -1,88 +1,85 @@
+from numpy import *
 from math import sqrt
-import numpy as np
 
+class Bases:
+    
+    def __init__(self, p1, p2, p3):
+                
+        self.p1 = array([p1[0], p1[1], p1[2]])
+        self.p2 = array([p2[0], p2[1], p2[2]])
+        self.p3 = array([p3[0], p3[1], p3[2]])
+        
+        print "p1: ", self.p1
+        print "p2: ", self.p2
+        print "p3: ", self.p3
 
-def angle_calculation(ap_axis, coil_axis):
-    # Function to calculate the angle (in degrees) between two given axis
-    ap_axis = np.array([ap_axis[0], ap_axis[1]])
-    coil_axis = np.array([float(coil_axis[0]), float(coil_axis[1])])
-    angle = np.rad2deg(np.arccos((np.dot(ap_axis, coil_axis))/(
-        np.linalg.norm(ap_axis)*np.linalg.norm(coil_axis))))
-
-    return float(angle)
-
-
-def base_creation(p1, p2, p3):
-    # Function to calculate the origin and matrix for coordinate system
-    # transformation.
-    # q: origin of coordinate system
-    # g1, g2, g3: orthogonal vectors of the coordinate system
-
-    p1 = np.array([p1[0], p1[1], p1[2]])
-    p2 = np.array([p2[0], p2[1], p2[2]])
-    p3 = np.array([p3[0], p3[1], p3[2]])
-
-    sub1 = p2 - p1
-    sub2 = p3 - p1
-    lamb = (sub1[0]*sub2[0]+sub1[1]*sub2[1]+sub1[2]*sub2[2])/np.dot(sub1, sub1)
-
-    q = p1 + lamb*sub1
-    g1 = p1 - q
-    g2 = p3 - q
-
-    if not g1.any():
-        g1 = p2 - q
-
-    g3 = np.cross(g2, g1)
-
-    g1 = g1/sqrt(np.dot(g1, g1))
-    g2 = g2/sqrt(np.dot(g2, g2))
-    g3 = g3/sqrt(np.dot(g3, g3))
-
-    m = np.matrix([[g1[0], g1[1], g1[2]],
-                   [g2[0], g2[1], g2[2]],
-                   [g3[0], g3[1], g3[2]]])
-
-    q.shape = (3, 1)
-    q = np.matrix(q.copy())
-    m_inv = m.I
-
-    print"M: ", m
-    print"q: ", q
-
-    return m, q, m_inv
-
-
-def flip_x(point):
-    # Function to flip the coordinates of a vector according to the X axis
-
-    # TODO: check if the Flip function is related to the X or Y axis
-    # Coronal Images dont require this transformation - 1 tested
-    # and for this case, at navigation, the z axis is inverted
+        self.sub1 = self.p2 - self.p1
+        self.sub2 = self.p3 - self.p1
+        
+    def Basecreation(self):
+        #g1
+        g1 = self.sub1
+        
+        #g2
+        lamb1 = g1[0]*self.sub2[0] + g1[1]*self.sub2[1] + g1[2]*self.sub2[2]
+        lamb2 = dot(g1, g1)
+        lamb = lamb1/lamb2
+        
+        #Ponto q    
+        q = self.p1 + lamb*self.sub1
+         
+        #g1 e g2 com origem em q   
+        g1 = self.p1 - q
+        g2 = self.p3 - q
+        
+        #testa se o g1 nao eh um vetor nulo
+        if g1.any() == False:
+            g1 = self.p2 - q
+            
+        #g3 - Produto vetorial NumPy
+        g3 = cross(g2, g1)
+        
+        #normalizacao dos vetores
+        g1 = g1/sqrt(lamb2)
+        g2 = g2/sqrt(dot(g2, g2))
+        g3 = g3/sqrt(dot(g3, g3))
+            
+        M = matrix([[g1[0],g1[1],g1[2]], [g2[0],g2[1],g2[2]], [g3[0],g3[1],g3[2]]])
+        q.shape = (3, 1)
+        q = matrix(q.copy())
+        print"M: ", M
+        print
+        print"q: ", q
+        print
+        Minv = M.I
+        
+        return M, q, Minv
+    
+def FlipX(point):
+        
+        point = matrix(point + (0,))
                
-    # It's necessary to multiply the z coordinate by (-1). Possibly
-    # because the origin of coordinate system of imagedata is
-    # located in superior left corner and the origin of VTK scene coordinate
-    #  system (polygonal surface) is in the interior left corner. Second
-    # possibility is the order of slice stacking
-
-    point = np.matrix(point + (0,))
-    point[0, 2] = -point[0, 2]
-
-    m_rot = np.matrix([[1.0, 0.0, 0.0, 0.0],
-                      [0.0, -1.0, 0.0, 0.0],
-                      [0.0, 0.0, -1.0, 0.0],
-                      [0.0, 0.0, 0.0, 1.0]])
-    m_trans = np.matrix([[1.0, 0, 0, -point[0, 0]],
-                        [0.0, 1.0, 0, -point[0, 1]],
-                        [0.0, 0.0, 1.0, -point[0, 2]],
-                        [0.0, 0.0, 0.0, 1.0]])
-    m_trans_return = np.matrix([[1.0, 0, 0, point[0, 0]],
+        #inverter o eixo z
+        ## possivel explicacaoo -- origem do eixo do imagedata esta no canto
+        ## superior esquerdo e origem da superfice eh no canto inferior esquerdo
+        ## ou a ordem de empilhamento das fatias
+        
+        point[0, 2] = -point[0, 2]
+        
+        #Flip em y
+        Mrot = matrix([[1.0, 0.0, 0.0, 0.0],
+                             [0.0, -1.0, 0.0, 0.0],
+                             [0.0, 0.0, -1.0, 0.0],
+                             [0.0, 0.0, 0.0, 1.0]])
+        Mtrans = matrix([[1.0, 0, 0, -point[0, 0]],
+                               [0.0, 1.0, 0, -point[0, 1]],
+                               [0.0, 0.0, 1.0, -point[0, 2]],
+                               [0.0, 0.0, 0.0, 1.0]])
+        Mtrans_return = matrix([[1.0, 0, 0, point[0, 0]],
                                [0.0, 1.0, 0, point[0, 1]],
                                [0.0, 0.0, 1.0, point[0, 2]],
                                [0.0, 0.0, 0.0, 1.0]])
         
-    point_rot = point*m_trans*m_rot*m_trans_return
-    x, y, z = point_rot.tolist()[0][:3]
-
-    return x, y, z
+        point_rot = point*Mtrans*Mrot*Mtrans_return
+        x, y, z  = point_rot.tolist()[0][:3]
+        return x, y, z
