+++ invesalius/data/imagedata_utils.py
@@ -531,68 +531,3 @@ def analyze2mmap(analyze):
 
     matrix.flush()
     return matrix, temp_file
-
-def nifti2mmap(nifti):
-    data = nifti.get_data()
-    header = nifti.get_header()
-    temp_file = tempfile.mktemp()
-
-#     # Sagital
-#     if header['orient'] == 2:
-#         print "Orientation Sagital"
-#         shape = tuple([data.shape[i] for i in (1, 2, 0)])
-#         matrix = numpy.memmap(temp_file, mode='w+', dtype=data.dtype, shape=shape)
-#         for n, slice in enumerate(data):
-#             matrix[:,:, n] = slice
-# 
-#     # Coronal
-#     elif header['orient'] == 1:
-#         print "Orientation coronal"
-#         shape = tuple([data.shape[i] for i in (1, 0, 2)])
-#         matrix = numpy.memmap(temp_file, mode='w+', dtype=data.dtype, shape=shape)
-#         for n, slice in enumerate(data):
-#             matrix[:,n,:] = slice
-# 
-#     # AXIAL
-#     elif header['orient'] == 0:
-#         print "no orientation"
-#         shape = tuple([data.shape[i] for i in (0, 1, 2)])
-#         matrix = numpy.memmap(temp_file, mode='w+', dtype=data.dtype, shape=shape)
-#         for n, slice in enumerate(data):
-#             matrix[n] = slice
-# 
-#     else:
-#         print "Orientation Sagital"
-#         shape = tuple([data.shape[i] for i in (1, 2, 0)])
-#         matrix = numpy.memmap(temp_file, mode='w+', dtype=data.dtype, shape=shape)
-#         for n, slice in enumerate(data):
-#             matrix[:,:, n] = slice
-
-    print " Orientation Axial"
-    shape = tuple([data.shape[i] for i in (0, 1, 2)])
-    matrix = numpy.memmap(temp_file, mode='w+', dtype=data.dtype, shape=shape)
-    data = imgnormalize(data)
-    dataswap = numpy.swapaxes(data, 2, 0)
-    dataswap[:] = dataswap[:, ::-1]
-    for n, slice in enumerate(dataswap):
-        print 'slice: ', slice
-        matrix[n] = slice
-    
-    matrix.flush()
-    return matrix, temp_file
-
-def imgnormalize(f, range=[0,255]):
-    f = numpy.asarray(f)
-    range = numpy.asarray(range)
-    faux = numpy.ravel(f).astype(float)
-    minimum = numpy.min(faux)
-    maximum = numpy.max(faux)
-    lower = range[0]
-    upper = range[1]
-    if minimum == maximum:
-        g = numpy.ones(f.shape)*(upper + lower) / 2.
-    else:
-        g = (faux-minimum)*(upper-lower) / (maximum-minimum) + lower
-    g = numpy.reshape(g, f.shape)
-    g = g.astype(f.dtype)
-    return g
