+++ invesalius/data/vtk_utils.py
@@ -142,7 +142,7 @@ class Text(object):
                                                     position[1])
 
     def GetPosition(self, position):
-        self.actor.GetPositionCoordinate().GetValue
+        self.actor.GetPositionCoordinate().GetValue()
 
     def SetJustificationToRight(self):
         self.property.SetJustificationToRight()
@@ -210,7 +210,7 @@ class TextZero(object):
                                                     position[1])
 
     def GetPosition(self, position):
-        self.actor.GetPositionCoordinate().GetValue
+        self.actor.GetPositionCoordinate().GetValue()
 
     def SetJustificationToRight(self):
         self.property.SetJustificationToRight()
